src/serverless/requirements.txt
boto3==1.16.1
botocore==1.19.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
f5-icontrol-rest==1.3.13
f5-sdk==3.0.21
idna==2.10
ipaddress==1.0.23
jmespath==0.10.0
netaddr==0.8.0
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
streaming-form-data==1.7.1
urllib3==1.25.11
validators==0.18.1